+++ lib/src/Widgets/wide_button.dart
@@ -5,7 +5,8 @@ class WideButton extends StatelessWidget {
       {Key? key,
       required,
       this.padding = 0.0,
-      this.height = 45,
+      this.height,
+      this.width,
       required this.onPressed,
       this.style,
       this.textColor})
@@ -13,7 +14,8 @@ class WideButton extends StatelessWidget {
 
   final String text;
   final double padding;
-  final double height;
+  final double? height;
+  final double? width;
   final ButtonStyle? style;
   final Color? textColor;
   final void Function() onPressed;
@@ -21,10 +23,11 @@ class WideButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SizedBox(
-      height: height,
-      width: MediaQuery.of(context).size.width <= 500
-          ? MediaQuery.of(context).size.width
-          : 350,
+      height: height ?? 45,
+      width: width ??
+          (MediaQuery.of(context).size.width <= 500
+              ? MediaQuery.of(context).size.width
+              : 350),
       child: Padding(
         padding: EdgeInsets.symmetric(horizontal: padding),
         child: ElevatedButton(